07-streaming/homework-7/pyproject.toml
[project]
name = "homework-7"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kafka-python>=2.3.0",
    "pandas>=3.0.2",
    "pyarrow>=23.0.1",
]
